Add `auto-toggle` command that turns light on/off when using a webcam
This adds a new `auto-toggle` command to the command line interface.
When using that command, `litra` will keep running and monitor the video
device files (`/dev/video*`) using inotify. Opening or closing such a
device will be tracked using a counter to determine if at least one
device is currently opened. If so, the light is turned on, otherwise
it's turned off.